move logic to metrics listener
