+++ internal/metrics/listener.go
@@ -12,7 +12,9 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/aws/aws-lambda-go/lambdacontext"
 	"runtime"
+	"strings"
 	"time"
 
 	"github.com/DataDog/datadog-lambda-go/internal/logger"
@@ -22,7 +24,7 @@ type (
 	// Listener implements wrapper.HandlerListener, injecting metrics into the context
 	Listener struct {
 		apiClient *APIClient
-		Config    *Config
+		config    *Config
 		processor Processor
 	}
 
@@ -60,19 +62,19 @@ func MakeListener(config Config) Listener {
 
 	return Listener{
 		apiClient: apiClient,
-		Config:    &config,
+		config:    &config,
 		processor: nil,
 	}
 }
 
 // HandlerStarted adds metrics service to the context
 func (l *Listener) HandlerStarted(ctx context.Context, msg json.RawMessage) context.Context {
-	if l.apiClient.apiKey == "" && l.Config.KMSAPIKey == "" && !l.Config.ShouldUseLogForwarder {
+	if l.apiClient.apiKey == "" && l.config.KMSAPIKey == "" && !l.config.ShouldUseLogForwarder {
 		logger.Error(fmt.Errorf("datadog api key isn't set, won't be able to send metrics"))
 	}
 
 	ts := MakeTimeService()
-	pr := MakeProcessor(ctx, l.apiClient, ts, l.Config.BatchInterval, l.Config.ShouldRetryOnFailure)
+	pr := MakeProcessor(ctx, l.apiClient, ts, l.config.BatchInterval, l.config.ShouldRetryOnFailure)
 	l.processor = pr
 
 	ctx = AddListener(ctx, l)
@@ -81,6 +83,7 @@ func (l *Listener) HandlerStarted(ctx context.Context, msg json.RawMessage) cont
 	l.apiClient.context = ctx
 
 	pr.StartProcessing()
+	l.submitEnhancedMetrics("invocations", ctx)
 
 	return ctx
 }
@@ -88,6 +91,9 @@ func (l *Listener) HandlerStarted(ctx context.Context, msg json.RawMessage) cont
 // HandlerFinished implemented as part of the wrapper.HandlerListener interface
 func (l *Listener) HandlerFinished(ctx context.Context) {
 	if l.processor != nil {
+		if ctx.Value("error") != nil {
+			l.submitEnhancedMetrics("errors", ctx)
+		}
 		l.processor.FinishProcessing()
 	}
 }
@@ -98,7 +104,7 @@ func (l *Listener) AddDistributionMetric(metric string, value float64, timestamp
 	// We add our own runtime tag to the metric for version tracking
 	tags = append(tags, getRuntimeTag())
 
-	if l.Config.ShouldUseLogForwarder {
+	if l.config.ShouldUseLogForwarder {
 		logger.Debug("sending metric via log forwarder")
 		unixTime := timestamp.Unix()
 		lm := logMetric{
@@ -125,7 +131,33 @@ func (l *Listener) AddDistributionMetric(metric string, value float64, timestamp
 	logger.Debug(fmt.Sprintf("adding metric \"%s\", with value %f", metric, value))
 	l.processor.AddMetric(&m)
 }
+
 func getRuntimeTag() string {
 	v := runtime.Version()
 	return fmt.Sprintf("dd_lambda_layer:datadog-%s", v)
+}
+
+func (l *Listener) submitEnhancedMetrics(metricName string, ctx context.Context) {
+	if l.config.EnhancedMetrics {
+		tags := getEnhancedMetricsTags(ctx)
+		l.AddDistributionMetric(fmt.Sprintf("aws.lambda.enhanced.%s", metricName), 1, time.Now(), tags...)
+	}
+}
+
+func getEnhancedMetricsTags(ctx context.Context) []string {
+	isColdStart := ctx.Value("cold_start")
+
+	if lc, ok := lambdacontext.FromContext(ctx); ok {
+		// ex: arn:aws:lambda:us-east-1:123497558138:function:golang-layer
+		splitArn := strings.Split(lc.InvokedFunctionArn, ":")
+
+		functionName := fmt.Sprintf("functionname:%s", lambdacontext.FunctionName)
+		region := fmt.Sprintf("region:%s", splitArn[3])
+		accountId := fmt.Sprintf("account_id:%s", splitArn[4])
+		memorySize := fmt.Sprintf("memorysize:%d", lambdacontext.MemoryLimitInMB)
+		coldStart := fmt.Sprintf("cold_start:%t", isColdStart.(bool))
+		return []string{functionName, region, accountId, memorySize, coldStart}
+	}
+	logger.Debug("could not retrieve the LambdaContext from Context")
+	return []string{}
 }
